delete_user

+++ server/routes/api/groups.routers.js
@@ -27,9 +27,10 @@ router.delete('/:groupId', async (req, res) => {
     if (req.session.user_id) {
       const currentUser = await User.findOne({
         where: { id: req.session.user_id },
+        include: { model: Role },
       });
       const currentGroup = await Group.findOne({ where: { id: groupId } });
-      if (currentGroup.teacher_id === req.session.user_id) {
+      if (currentUser.Role.title === 'Учитель' && currentGroup.teacher_id === req.session.user_id) {
         const result = await Group.destroy({ where: { id: groupId } });
         if (result > 0) {
           res.status(200).json(+groupId);
@@ -47,7 +48,6 @@ router.delete('/:groupId', async (req, res) => {
     res.status(400).json({ message: 'Postman не пройдет' });
   }
 });
-
 
 router.post('/', async (req, res) => {
   try {
@@ -70,5 +70,25 @@ router.post('/', async (req, res) => {
     res.json(message);
   }
 });
+
+// router.put('/:groupId', async (req, res) => {
+//   const { groupId } = req.params;
+//   const { title } = req.body;
+//   try {
+//     const group = await Group.findOne({ where: { id: groupId } });
+//     if (oneModule.user_id === req.session.user_id) {
+//       oneModule.title = title;
+//       oneModule.category = categoryId;
+//       oneModule.save();
+//       res.json(oneModule);
+//       return;
+//     } else {
+//       res.json({ message: 'Не отработал' });
+//       return;
+//     }
+//   } catch ({ message }) {
+//     res.json({ message });
+//   }
+// });
 
 module.exports = router;
